permissions - delete
git-svn-id: http://www.pimcore.org/svn/pimcore/private/core/trunk@4746 1f8fe7d8-47f0-464c-8d0a-336f4953ab05

+++ pimcore/modules/admin/controllers/AssetController.php
@@ -387,7 +387,9 @@ class Admin_AssetController extends Pimcore_Controller_Action_Admin
             $deletedItems = array();
             foreach ($assets as $asset) {
                 $deletedItems[] = $asset->getFullPath();
-                $asset->delete();
+                if ($asset->isAllowed("delete")) {
+                    $asset->delete();
+                }
             }
 
             $this->_helper->json(array("success" => true, "deleted" => $deletedItems));
