+++ pimcore/modules/admin/controllers/DocumentController.php
@@ -257,7 +257,9 @@ class Admin_DocumentController extends Pimcore_Controller_Action_Admin
             $deletedItems = array();
             foreach ($documents as $document) {
                 $deletedItems[] = $document->getFullPath();
-                $document->delete();
+                if ($document->isAllowed("delete")) {
+                    $document->delete();
+                }
             }
 
             $this->_helper->json(array("success" => true, "deleted" => $deletedItems));
