+++ pimcore/modules/admin/controllers/ObjectController.php
@@ -767,7 +767,9 @@ class Admin_ObjectController extends Pimcore_Controller_Action_Admin
             $deletedItems = array();
             foreach ($objects as $object) {
                 $deletedItems[] = $object->getFullPath();
-                $object->delete();
+                if ($object->isAllowed("delete")) {
+                    $object->delete();
+                }
             }
 
             $this->_helper->json(array("success" => true, "deleted" => $deletedItems));
